requirements.txt
python-docx
numpy==2.1.*
Flask==3.1.1
flask_login==0.6.3
flask_sqlalchemy==3.1.1
ip2geotools==0.1.6
openai==0.28.0
pdf2image==1.17.0
Pillow==10.2.0
pytesseract==0.3.13
SQLAlchemy==2.0.41
Werkzeug==3.1.3
gunicorn==22.0.0
spacy>=3.7,<3.9
transformers==4.53.0
accelerate>=0.34
llama-cpp-python==0.2.87
psycopg2-binary>=2.9,<3
cryptography>=41.0.0
typed-ast==1.5.5

google-cloud-vision>=3.7.0
